Rename transducer class
